+++ utils-core/src/main/scala/org/tupol/spark/utils.scala
@@ -24,10 +24,6 @@ SOFTWARE.
 package org.tupol.spark
 
 import java.net.{ URI, URL }
-import java.sql.Timestamp
-import java.time.LocalDateTime
-import org.json4s.JsonAST.JString
-import org.json4s.{ CustomSerializer, Serializer }
 import org.tupol.utils.Bracket
 import org.tupol.utils.implicits._
 
@@ -36,36 +32,6 @@ import scala.util.{ Failure, Try }
 
 /** A few common functions that might be useful. */
 package object utils extends Logging {
-
-  /**
-   * This is a small and probably wrong conversion to JSON format.
-   *
-   * Besides the basic conversion, this also serializes the LocalDateFormat
-   *
-   * @param input input to be converted to JSON format
-   * @return
-   */
-  def toJson(input: AnyRef) = {
-    //TODO Find a nicer more comprehensive solution
-    import org.json4s.NoTypeHints
-    import org.json4s.jackson.Serialization
-
-    implicit val formats = Serialization.formats(NoTypeHints) ++ TimeSerializers
-    Serialization.write(input)
-  }
-
-  /** Serializers for Time types that use commonly */
-  val TimeSerializers: Seq[Serializer[_]] = {
-    /** Serializer / deserializer for LocalDateFormat */
-    case object LDTSerializer extends CustomSerializer[LocalDateTime](format => (
-      { case JString(s) => LocalDateTime.parse(s) },
-      { case ldt: LocalDateTime => JString(ldt.toString) }))
-    /** Serializer / deserializer for Timestamp */
-    case object SqlTimestampSerializer extends CustomSerializer[Timestamp](format => (
-      { case JString(ts) => Timestamp.valueOf(LocalDateTime.parse(ts)) },
-      { case ts: Timestamp => JString(ts.toLocalDateTime.toString) }))
-    Seq(LDTSerializer, SqlTimestampSerializer)
-  }
 
   /**
    * Try loading a text resource from a given path, whether it is local, from a given URL or URI or from the classpath.