+++ utils-io-pureconfig/src/test/scala/org/tupol/spark/io/pureconf/FormatAwareDataSourceConfigurationSpec.scala
@@ -35,7 +35,7 @@ class FormatAwareDataSourceConfigurationSpec extends AnyFunSuite with Matchers w
 
     val configStr =
       """
-        |input.format="delta"
+        |input.format="my-format"
         |input.options={
         |  path: "INPUT_PATH"
         |}
@@ -43,7 +43,7 @@ class FormatAwareDataSourceConfigurationSpec extends AnyFunSuite with Matchers w
     val config = ConfigFactory.parseString(configStr)
 
     val expected = GenericSourceConfiguration(
-      FormatType.Custom("delta"),
+      FormatType.Custom("my-format"),
       options = Map("path" -> "INPUT_PATH"),
       schema = None)